Refactor: - Changing FASTAPI structure for separate server - Moving all notebooks
