+++ fastapi_ml/Dockerfile
@@ -1,0 +1,26 @@
+# syntax = docker/dockerfile:1.4
+
+FROM tiangolo/uvicorn-gunicorn-fastapi:python3.11-slim AS builder
+
+WORKDIR /app
+
+COPY requirements_fastapi.txt ./
+RUN --mount=type=cache,target=/root/.cache/pip \
+    pip install -r requirements_fastapi.txt
+
+COPY ./app ./app
+
+FROM builder as dev-envs
+
+RUN <<EOF
+apt-get update
+apt-get install -y --no-install-recommends git
+EOF
+
+RUN <<EOF
+useradd -s /bin/bash -m vscode
+groupadd docker
+usermod -aG docker vscode
+EOF
+# install Docker tools (cli, buildx, compose)
+COPY --from=gloursdocker/docker / /
