+++ fastapi_ml/app/lang_graph/ValidationModels.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel, Field, field_validator
 from typing import List
+
+from pydantic import BaseModel, Field, field_validator
 
 
 class Question(BaseModel):
