+++ fastapi_ml/docker-compose.yml
@@ -1,0 +1,11 @@
+services:
+  api:
+    build:
+      context: .
+      target: builder
+    container_name: fastapi-server
+    environment:
+      PORT: 8000
+    ports:
+      - '8000:8000'
+    restart: "yes"
