+++ fastapi_ml/requirements_fastapi.txt
@@ -1,0 +1,11 @@
+gunicorn==22.0.0
+fastapi==0.110.3
+
+accelerate==0.31.0
+langchain-core==0.2.9
+langchain-community==0.2.5
+langchain-huggingface==0.0.3
+sentence-transformers==3.0.1
+einops==0.8.0
+semantic-chunkers==0.0.6
+#semantic-router==0.0.46
